refactor(apple): use scoped resolvers on macOS
