+++ swift/apple/FirezoneNetworkExtension/Adapter.swift
@@ -56,7 +56,7 @@ actor Adapter {
   private var startContinuation: CheckedContinuation<Void, Error>?
 
   /// Used for finding system DNS resolvers when network conditions have changed.
-  private let systemConfigurationResolvers: SystemConfigurationResolvers
+  private let scopedResolvers: ScopedResolvers
 
   /// Remembers the last _relevant_ path update.
   /// A path update is considered relevant if certain properties change that require us to reset connlib's
@@ -73,10 +73,8 @@ actor Adapter {
   /// Apple doesn't give us very much info in this callback, so we don't know which of the
   /// events above triggered the callback.
   ///
-  /// On macOS, we use the SystemConfiguration framework to read the system's DNS resolvers directly.
-  ///
-  /// On iOS, we use dlsym to access the `dns_configuration_copy` function which
-  /// returns scoped resolvers that aren't shadowed by our tunnel's DNS settings.
+  /// We use dlsym to access the `dns_configuration_copy` function which returns scoped
+  /// resolvers that aren't shadowed by our tunnel's DNS settings.
   ///
   /// See the following issues for background:
   /// - https://github.com/firezone/firezone/issues/3302
@@ -135,7 +133,7 @@ actor Adapter {
     accountSlug: String,
     internetResourceEnabled: Bool,
     providerCommandSender: Sender<ProviderCommand>
-  ) throws {
+  ) {
     self.apiURL = apiURL
     self.token = token
     self.deviceId = deviceId
@@ -144,7 +142,7 @@ actor Adapter {
     self.internetResourceEnabled = internetResourceEnabled
     self.providerCommandSender = providerCommandSender
     self.unreachableResources = []
-    self.systemConfigurationResolvers = try SystemConfigurationResolvers()
+    self.scopedResolvers = ScopedResolvers()
 
     // Start log cleanup immediately - doesn't depend on tunnel being connected
     providerCommandSender.send(.startLogCleanupTask)
@@ -451,10 +449,10 @@ actor Adapter {
 
   private func setSystemDefaultResolvers(_ path: Network.NWPath) async {
     // Step 1: Get system default resolvers
-    let resolvers = self.systemConfigurationResolvers.getDefaultDNSServers(
+    let resolvers = self.scopedResolvers.getDefaultDNSServers(
       interfaceName: path.availableInterfaces.first?.name)
 
-    // Step 2: Validate and strip scope suffixes
+    // Step 2: Validate addresses and filter out sentinel ranges
     var parsedResolvers: [String] = []
 
     for stringAddress in resolvers {
